moving sorting upstream app versions into specialized package
