+++ pkg/helm/helm.go
@@ -9,11 +9,11 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/Masterminds/semver"
 	"github.com/go-git/go-billy/v5"
 	"github.com/rancher/charts-build-scripts/pkg/filesystem"
 	"github.com/rancher/charts-build-scripts/pkg/logger"
 	"github.com/rancher/charts-build-scripts/pkg/path"
+	"github.com/rancher/charts-build-scripts/pkg/util"
 	helmRepo "helm.sh/helm/v3/pkg/repo"
 )
 
@@ -198,66 +198,7 @@ func OpenIndexYaml(ctx context.Context, rootFs billy.Filesystem) (*helmRepo.Inde
 func SortVersions(index *helmRepo.IndexFile) {
 	for _, versions := range index.Entries {
 		sort.Slice(versions, func(i, j int) bool {
-			return compareVersions(versions[i].Version, versions[j].Version)
+			return util.SortUpstreamAppVersions(versions[i].Version, versions[j].Version)
 		})
 	}
-}
-
-// compareVersions compares two version strings for sorting
-// Returns true if versionA should come before versionB (descending order)
-// Handles alpha, beta, rc, and stable versions
-func compareVersions(versionA, versionB string) bool {
-	// Split versions into upstream and chart parts
-	// Format: <upstream_version>+up<chart_version>
-	upstreamA, chartA := splitUpstreamChart(versionA)
-	upstreamB, chartB := splitUpstreamChart(versionB)
-
-	// Parse upstream versions using semver
-	semverUpstreamA, errA := semver.NewVersion(upstreamA)
-	semverUpstreamB, errB := semver.NewVersion(upstreamB)
-
-	if errA != nil {
-		return false // push invalid to end
-	}
-	if errB != nil {
-		return true // push invalid to end
-	}
-
-	// If upstream versions are different, use semver comparison (descending)
-	if !semverUpstreamA.Equal(semverUpstreamB) {
-		return semverUpstreamA.GreaterThan(semverUpstreamB)
-	}
-
-	// Same upstream version - compare chart versions (the part after +up)
-	if chartA != "" && chartB != "" {
-		semverChartA, errA := semver.NewVersion(chartA)
-		semverChartB, errB := semver.NewVersion(chartB)
-
-		if errA != nil {
-			return false // push invalid to end
-		}
-		if errB != nil {
-			return true // push invalid to end
-		}
-
-		// Compare chart versions (descending)
-		// This automatically handles prereleases correctly per semver spec
-		if !semverChartA.Equal(semverChartB) {
-			return semverChartA.GreaterThan(semverChartB)
-		}
-	}
-
-	// Versions are equal
-	return false
-}
-
-// splitUpstreamChart splits a version string into upstream and chart parts
-// Example: "109.0.1+up0.10.4-rc.1" returns ("109.0.1", "0.10.4-rc.1")
-// Example: "109.0.1" returns ("109.0.1", "")
-func splitUpstreamChart(version string) (upstream, chart string) {
-	parts := strings.Split(version, "+up")
-	if len(parts) == 2 {
-		return parts[0], parts[1]
-	}
-	return version, ""
 }
